refactor: remove unused files and tests; update testing configuration and utilities
