+++ pytest.ini
@@ -1,0 +1,10 @@
+[pytest]
+DJANGO_SETTINGS_MODULE = app.settings.development
+python_files = tests.py test_*.py *_tests.py *_test.py
+filterwarnings =
+    ignore::DeprecationWarning
+    ignore::UserWarning
+markers =
+    unit: Unit tests
+    integration: Integration tests
+    slow: Slow running tests
